Move `compute_additions` functions into wallet.
