+++ chia/wallet/singleton.py
@@ -14,7 +14,7 @@ from chia_rs.sized_bytes import bytes32
 from chia.types.blockchain_format.coin import Coin
 from chia.types.blockchain_format.program import Program
 from chia.types.blockchain_format.serialized_program import SerializedProgram
-from chia.types.coin_spend import compute_additions
+from chia.wallet.util.compute_additions import compute_additions
 from chia.wallet.util.curry_and_treehash import (
     calculate_hash_of_quoted_mod_hash,
     curry_and_treehash,
